requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
psycopg[binary]==3.2.10
python-dotenv==1.0.0
reportlab==4.0.4
matplotlib==3.8.2
requests==2.31.0

# Development dependencies
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0
